server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
torch==2.1.2
transformers==4.48.3
faiss-cpu==1.11.0
numpy==1.25.2
pillow==11.1.0
scipy==1.15.1
scikit-learn==1.2.2
scikit-image
tqdm==4.67.1
protobuf==6.30.2
datasets==3.3.2
sentencepiece==0.2.0
accelerate==1.3.0
pydantic-settings==2.11.0
psutil
pytest
requests
pyyaml
# SAM 3 — install separately: cd server/sam3 && pip install -e .
# Or run: bash scripts/setup_models.sh
